Added shared StorPools
Shared storage pools are handled by the FreeSpaceManager. One
FreeSpaceManager can handle one or more storpools.

The user can specify a --shared-space name for a StorPool. All
StorPools with the same shared-space-name have the same
FreeSpaceManager instance.

When the Controller creates a Volume, that Volume is in an "adding"
state for the FreeSpaceManager. In this state, the FreeSpaceManager
still knows the last reported exact free space, but is also able to
calculate an estimation including the Volume(s) currently in the
"adding" state. These two values can differ as some StorageDrivers
(like LVM) have an extent size, which will round up the requested
size.

Once the Satellite finished deploying the Volume(s), the reported
success event to the Controller contains the new exact free space.
With this event the FreeSpaceManager removes the Volume from its
"adding" state.

Removing a Volume has the same procedure with a seperate
state-tracking.

+++ server/src/main/java/com/linbit/linstor/VolumeData.java
@@ -242,7 +242,17 @@ public class VolumeData extends BaseTransactionObject implements Volume
     {
         checkDeleted();
         resource.getObjProt().requireAccess(accCtx, AccessType.CHANGE);
+
+        StorPool oldStorPool = storPool.get();
+        if (oldStorPool != null)
+        {
+            if (oldStorPool.containsVolume(accCtx, this))
+            {
+                oldStorPool.getFreeSpaceManager().removingVolume(accCtx, this);
+            }
+        }
         storPool.set(storPoolRef);
+        storPoolRef.putVolume(accCtx, this);
     }
 
     @Override
@@ -305,6 +315,7 @@ public class VolumeData extends BaseTransactionObject implements Volume
         checkDeleted();
         resource.getObjProt().requireAccess(accCtx, AccessType.USE);
         getFlags().enableFlags(accCtx, Volume.VlmFlags.DELETE);
+        storPool.get().getFreeSpaceManager().removingVolume(accCtx, this);
     }
 
     @Override
@@ -313,7 +324,7 @@ public class VolumeData extends BaseTransactionObject implements Volume
         checkDeleted();
         resource.getObjProt().requireAccess(accCtx, AccessType.VIEW);
 
-        return realSize.get() == null;
+        return realSize.get() != null;
     }
 
     @Override
@@ -366,7 +377,12 @@ public class VolumeData extends BaseTransactionObject implements Volume
             }
 
             ((ResourceData) resource).removeVolume(accCtx, this);
-            storPool.get().removeVolume(accCtx, this);
+            if (storPool.get().containsVolume(accCtx, this))
+            {
+                throw new ImplementationError(
+                    "A volume was not removed from its storPool (which includes an update to the freeSpaceMgr)"
+                );
+            }
             ((VolumeDefinitionData) volumeDfn).removeVolume(accCtx, this);
 
             volumeProps.delete();
